requirements.txt
# Core dependencies
numpy==1.23.5
pandas==1.5.3
matplotlib>=3.4.0
scikit-learn>=1.0.0
Pillow>=9.4.0
requests>=2.25.0
python-dotenv>=1.0.0
transformers>=4.0.0
rapidfuzz>=1.6.0
pytesseract>=0.3.9
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.15.1; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.15.1; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.15.1; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
torch>=2.0.1
torchvision>=0.15.2
sentence-transformers>=2.0.0

# API documentation
fastapi>=0.115.6
uvicorn>=0.22.0
python-multipart>=0.0.6

# NLP tools
openai>=0.27.0
opencv-python-headless==4.7.0.72

# Cloud integration
google-cloud-storage>=2.0.0

# Development tools
pytest>=6.0.0
pylint>=2.8.0
black>=22.0.0
